requirements.txt
streamlit==1.37.1
pandas==2.2.2
numpy==1.26.4
openpyxl==3.1.5
xlsxwriter==3.2.0
pyarrow==16.1.0
streamlit-authenticator==0.4.2
PyYAML==6.0.2
